V0.2.6

+++ test/lib.js
@@ -7,20 +7,17 @@ var should = chai.should();  // Using Should style
 
 const fs = require('fs')
 
-const db = ':memory:'
+const async = false;
+const db = `./example_async_${async}.db`
 const table = 'admin'
 
-const async = false;
 
-describe('Library (sync)', function() {
-
-  if (fs.existsSync(db)) {
-    fs.unlinkSync(db);
-  }
+describe(`Library (async: ${async})`, function() {
+  if (fs.existsSync(db)) fs.unlinkSync(db);
 
   let connection = new Connection(db);
 
-  describe('Connection', function() {
+  describe('Create', function() {
     it('should return false when create() is failed', function() {
       let stmt = connection.createStatement();
       let query = `CREATE TABLE ${table}(idx INTEGER PRIMARY KEY AUTOINCREMENT, passwd TEXT, dump BLOB, date DATETIME);`;
@@ -30,6 +27,7 @@ describe('Library (sync)', function() {
       })
       .then(res => {
         assert.equal(res, true);
+        assert.equal(fs.existsSync(db), true);
       })
       .catch(err => {
         console.log(err);
@@ -82,7 +80,45 @@ describe('Library (sync)', function() {
       });
     });
 
-    // PRAGMA table_info(admin)
+    it('should return false when create() is failed', function() {
+      let stmt = connection.createStatement();
+      let query = `PRAGMA table_info(${table});`;
+      stmt.executeQuery({
+        query: query,
+        async: async,
+      })
+      .then(res => {
+        assert.equal(res.next(), true);
+        let obj = res.obj;
+        assert.equal(obj.hasOwnProperty('name'), true);
+        assert.equal(obj.name, 'idx');
+        assert.equal(obj.type, 'INTEGER');
+
+        assert.equal(res.next(), true);
+        obj = res.obj;
+        assert.equal(obj.hasOwnProperty('name'), true);
+        assert.equal(obj.name, 'passwd');
+        assert.equal(obj.type, 'TEXT');
+
+        assert.equal(res.next(), true);
+        obj = res.obj;
+        assert.equal(obj.hasOwnProperty('name'), true);
+        assert.equal(obj.name, 'dump');
+        assert.equal(obj.type, 'BLOB');
+
+        assert.equal(res.next(), true);
+        obj = res.obj;
+        assert.equal(obj.hasOwnProperty('name'), true);
+        assert.equal(obj.name, 'date');
+        assert.equal(obj.type, 'DATETIME');
+      })
+      .catch(err => {
+        console.log(err);
+      });
+    });
+  });
+
+  describe('Insert', function() {
     it('should return false when insert() is failed', function() {
       let stmt = connection.prepareStatement(`INSERT INTO ${table} (passwd, dump, date) VALUES(?,?,datetime(\'now\',\'localtime\'));`);
       stmt.setString({
@@ -113,7 +149,9 @@ describe('Library (sync)', function() {
         console.log(err);
       });
     });
+  });
 
+  describe('Select', function() {
     it('should return error when select() is failed', function() {
       let stmt = connection.createStatement();
       let query = `SELECT idx, passwd, dump, date FROM ${table}`;
@@ -140,7 +178,9 @@ describe('Library (sync)', function() {
         console.log(err);
       });
     });
+  });
 
+  describe('Update', function() {
     it('should return false when update() is failed', function() {
       let stmt = connection.prepareStatement(`UPDATE ${table} set passwd=?, date=datetime(\'now\',\'localtime\') WHERE idx=1;`);
       stmt.setString({
@@ -175,7 +215,9 @@ describe('Library (sync)', function() {
         console.log(err);
       });
     });
+  });
 
+  describe('Delete', function() {
     it('should return false when delete() is failed', function() {
       let stmt = connection.prepareStatement(`DELETE FROM ${table} WHERE idx=?;`);
       let id = 2;
@@ -208,6 +250,7 @@ describe('Library (sync)', function() {
     });
   });
 
+  let count = 1000;
   describe('Performance', function() {
     it('should return false when insert() is failed', function() {
       for(var i = 0; i < 1000; i++) {
@@ -217,7 +260,9 @@ describe('Library (sync)', function() {
           value: 'admin_passwd',
         });
 
-        stmt.execute(async)
+        stmt.execute({
+          async: async,
+        })
         .then(res => {
           assert.equal(res, true);
         })
@@ -243,13 +288,13 @@ describe('Library (sync)', function() {
     });
 
     it('should return false when insert() is failed', function() {
-      for(var i = 0; i < 1000; i++) {
+      for(var i = 0; i < count; i++) {
         let stmt = connection.createStatement();
         let passwd = `admin_passwd${i}`;
         let query = `INSERT INTO ${table} (passwd, date) VALUES(\'${passwd}\',datetime(\'now\',\'localtime\'));`;
         stmt.execute({
           query: query,
-          async: async
+          async: async,
         })
         .then(res => {
           assert.equal(res, true);
@@ -269,7 +314,7 @@ describe('Library (sync)', function() {
       })
       .then(res => {
         res.next();
-        assert.equal(res.data[0], 1000);
+        assert.equal(res.data[0], count);
       })
       .catch(err => {
         console.log(err);
